requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
babel==2.17.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
filetype==1.2.0
flask==3.1.0
flask-babel==4.0.0
flask-login==0.6.3
flask-oauthlib==0.9.6
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-core==0.3.35
langchain-google-genai==2.0.9
langchain-text-splitters==0.3.6
langsmith==0.3.8
markdown==3.7
markupsafe==3.0.2
multidict==6.1.0
numpy==2.2.3
oauthlib==2.1.0
orjson==3.10.15
packaging==24.2
pillow==11.1.0
polib==1.2.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
pyparsing==3.2.1
pytz==2025.1
pyyaml==6.0.2
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==1.1.0
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
sqlalchemy==2.0.38
tenacity==9.0.0
tqdm==4.67.1
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
werkzeug==3.1.3
yarl==1.18.3
zstandard==0.23.0
